Creando logica de cupon y sincupon

+++ TP1J/src/src/Garantia.java
@@ -3,7 +3,7 @@ package src;
 public class Garantia extends Servicio {
 
 	@Override
-	int modificarPrecio(int precio) {
+	double modificarPrecio(double precio) {
 		return (precio*1.1);
 	}
