+++ TP1J/src/src/Servicio.java
@@ -2,6 +2,6 @@ package src;
 
 abstract class Servicio {
 	
-	abstract int modificarPrecio(int precio);
+	abstract double modificarPrecio(double precio);
 
 }
